implemented HarvesterController - verifies request, installs harvester. added harvesters to empire game data implemented frontend components to show harvesters, including info modal implemented turn processing for harvesters - building harvesters, resource production.

+++ database/migrations/2020_11_04_202059_create_harvesters_table.php
@@ -22,7 +22,7 @@ class CreateHarvestersTable extends Migration
             $table->uuid('game_id');
             $table->uuid('player_id');
             $table->enum('resource_type', array_keys(config('rules.harvesters')));
-            $table->smallInteger('production');
+            $table->float('production', 4,2);
             $table->tinyInteger('until_complete');
             $table->foreign('planet_id')->references('id')->on('planets')
                 ->onDelete('cascade');
